+++ githubCard/src/App.js
@@ -1,5 +1,7 @@
 import React from 'react';
 import './App.css';
+import Search from './Search'
+import UserCard from './UserCard'
 
 class App extends React.Component {
   constructor() {
@@ -13,13 +15,14 @@ class App extends React.Component {
   }
 
   changeUserName = (userName) => {
-    this.setState({ userName });
+    this.setState( {userName: userName });
   }
 
   fetchUser = () => {
     fetch(`https://api.github.com/users/${this.state.userName}`)
       .then(res => res.json())
-      .then(data => this.setState({ user: data }));
+      .then(data => this.setState({ user: data }))
+      
   }
 
   fetchFollowers = () => {
@@ -58,52 +61,6 @@ class App extends React.Component {
   }
 }
 
-class Search extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      search: "",
-    };
-  }
 
-  handleChange = event => {
-    this.setState({ [event.target.name]: event.target.value });
-  }
-
-  handleSubmit = event => {
-    event.preventDefault();
-    this.props.changeUserName(this.state.search);
-    this.setState({ search: "" });
-  }
-
-  render() {
-    return (
-      <form onSubmit={this.handleSubmit} >
-        <input type="text"
-               name="search"
-               placeholder="search"
-               value={this.state.search}
-               onChange={this.handleChange} />
-        <button type="submit">Search for a User</button>
-      </form>
-    );
-  }
-}
-
-function UserCard(props) {
-  return (
-    <div className="card">
-      <h2>{props.user.login}</h2>
-      <p className="location">{props.user.location}</p>
-      
-      <div>
-        <h4>Friends: </h4>
-        {props.followers.map(follower => (
-          <div className="name" key={follower.id}>{follower.login}</div>
-        ))}
-      </div>
-    </div>
-  );
-}
 
 export default App;
